requirements-inf.txt
onnxruntime==1.20.1
Pillow==11.0.0
rasterio==1.4.2
numpy==2.1.3
pystac==1.13.0
